.tekton: enabling hermetic builds
This will enforce Konflux to prefetch the dependencies defined in
the lock.yaml files with [1]. Then during the build, Konflux will
1. inject the repositories where the deps are stored, 2. configure
the clients to pull the deps from there, 3. build without network.

As rpm is still not fully supported [2], we have to enable
`dev-package-managers` for now in the pipeline.

All specific files enabling hermetic builds are located in the
`ci/hermetic/` folder. You can find the helper scripts that automate
the process of generating the lock YAML files, replacing the manual
steps. The automation streamlines the workflow, reduces the chance
of human error, and ensures consistency in the generated lock files.
More details can be found in the updated README.

This required adaptations to `build.sh` and the Dockerfile to support
both hermetic and non-hermetic build processes.

[1] https://github.com/konflux-ci/build-definitions/tree/main/task/prefetch-dependencies-oci-ta/0.2
[2] https://github.com/hermetoproject/hermeto?tab=readme-ov-file#package-managers