tests/containers: switch to standard base image
In 'rpms.in.yaml' file, we configured the rpm lockfile CLI tool to
extract installed packages from the main Dockerfile which use the
standard base image of Fedora. For consistency, we have to rbase
all the container images of the project on the same image.
Otherwise, we may end up with missing packages during the build
(i.e tzdata is available in standard image, but not in minimal
one, making the targetcli fail as tzdata was not add to the lock
YAML file).